rework doc structure
